- updates in prep for 3.5.2

+++ mk/linux/makedata.sh
@@ -50,7 +50,7 @@ svn export --force "$CURRENTDIR/../../source/masterserver/flags" "$RELEASEDIR/da
 cp -p "$CURRENTDIR/../../data/glest_game/CMakeLists.txt" $RELEASEDIR
 
 echo "creating $PACKAGE"
-rm "release/$PACKAGE"
+[[ -f "$release/$PACKAGE" ]] && rm "release/$PACKAGE"
 #tar cJf "release/$PACKAGE" -C "$CURRENTDIR/release/" "$RELEASENAME-$VERSION"
 tar -cf - -C "$CURRENTDIR/release/$RELEASENAME-$VERSION/" "megaglest-$VERSION" | xz -9e > release/$PACKAGE
 # 7z a -mmt -mx=9 -ms=on -mhc=on "release/$PACKAGE" "$CURRENTDIR/release/$RELEASENAME-$VERSION"
